Keeps a static trace of dynamic deps
The idea is to assume that, by default, a task will have dynamic
dependencies and that the `no_dynamic_deps` tag transforms the arrow to
explicitly tell it that it has no dynamic dependencies.
This enables a static observation to be made about whether or not
dynamic dependencies should be tested.