Add easier function for handling static deps
As mentioned by @xvw, in most cases, we build a blog with minimal
dynamic dependencies. Therefore, we have added an action facilitating
the creation of a "static with dependency" file. This enables more
efficient and streamlined construction processes, aligning with the
overarching principle of reducing unnecessary complexity. This
enhancement serves to optimize the blog's performance and maintenance,
ensuring smoother operations and better scalability in the long run.

Co-authored-by: xvw <xaviervdw@gmail.com>